// src/types/currency.ts
export type CurrencyCode = "USD" | "CLP" | "BRL"

export interface OpenExchangeRatesResponse {
	disclaimer: "Usage subject to terms: https://openexchangerates.org/terms"
	license: "https://openexchangerates.org/license"
	timestamp: 1729551600
	base: "USD"
	rates: {
		AED: 3.673
		AFN: 65.556031
		ALL: 91.063155
		AMD: 386.745404
		ANG: 1.802151
		AOA: 910.833667
		ARS: 983.2682
		AUD: 1.50285
		AWG: 1.8025
		AZN: 1.7
		BAM: 1.802888
		BBD: 2
		BDT: 119.495013
		BGN: 1.808425
		BHD: 0.376969
		BIF: 2897.463456
		BMD: 1
		BND: 1.312595
		BOB: 6.923821
		BRL: 5.6896
		BSD: 1
		BTC: 0.00001484941
		BTN: 84.050604
		BWP: 13.339785
		BYN: 3.272295
		BZD: 2.0153
		CAD: 1.383421
		CDF: 2842.785693
		CHF: 0.865571
		CLF: 0.034504
		CLP: 952.25
		CNH: 7.13843
		CNY: 7.1189
		COP: 4285.079499
		CRC: 514.189183
		CUC: 1
		CUP: 25.75
		CVE: 101.906449
		CZK: 23.366981
		DJF: 177.762377
		DKK: 6.895345
		DOP: 60.264303
		DZD: 133.756124
		EGP: 48.6691
		ERN: 15
		ETB: 119.561718
		EUR: 0.924608
		FJD: 2.2392
		FKP: 0.770245
		GBP: 0.770245
		GEL: 2.72
		GGP: 0.770245
		GHS: 16.032147
		GIP: 0.770245
		GMD: 70
		GNF: 8625.317442
		GTQ: 7.732194
		GYD: 209.078753
		HKD: 7.773897
		HNL: 25.05
		HRK: 6.966305
		HTG: 131.615704
		HUF: 371.157553
		IDR: 15560.935936
		ILS: 3.780925
		IMP: 0.770245
		INR: 84.07445
		IQD: 1309.92089
		IRR: 42102.5
		ISK: 137.86
		JEP: 0.770245
		JMD: 158.902628
		JOD: 0.709
		JPY: 150.6235
		KES: 128.994286
		KGS: 85.5
		KHR: 4058.397336
		KMF: 455.150003
		KPW: 900
		KRW: 1377.505579
		KWD: 0.306485
		KYD: 0.833218
		KZT: 482.169905
		LAK: 21948.385651
		LBP: 89552.734108
		LKR: 293.163724
		LRD: 192.292091
		LSL: 17.621084
		LYD: 4.811129
		MAD: 9.909067
		MDL: 17.727447
		MGA: 4605.868185
		MKD: 56.893617
		MMK: 2098
		MNT: 3398
		MOP: 8.004148
		MRU: 39.695798
		MUR: 45.899997
		MVR: 15.36
		MWK: 1734.319519
		MXN: 19.968959
		MYR: 4.309
		MZN: 63.850001
		NAD: 17.621084
		NGN: 1636.521416
		NIO: 36.77934
		NOK: 10.952155
		NPR: 134.492595
		NZD: 1.660261
		OMR: 0.38493
		PAB: 1
		PEN: 3.754594
		PGK: 3.957877
		PHP: 57.656002
		PKR: 277.678127
		PLN: 3.994438
		PYG: 7921.221624
		QAR: 3.641771
		RON: 4.5984
		RSD: 108.212825
		RUB: 96.802286
		RWF: 1351.613728
		SAR: 3.756322
		SBD: 8.299327
		SCR: 13.619658
		SDG: 601.5
		SEK: 10.563239
		SGD: 1.31669
		SHP: 0.770245
		SLL: 20969.5
		SOS: 571.217617
		SRD: 33.215
		SSP: 130.26
		STD: 22281.8
		STN: 22.794763
		SVC: 8.748957
		SYP: 2512.53
		SZL: 17.618098
		THB: 33.507621
		TJS: 10.652933
		TMT: 3.5
		TND: 3.106065
		TOP: 2.370898
		TRY: 34.251302
		TTD: 6.785364
		TWD: 32.066001
		TZS: 2724.730924
		UAH: 41.309706
		UGX: 3665.145572
		USD: 1
		UYU: 41.568799
		UZS: 12817.443582
		VES: 39.132497
		VND: 25303.658306
		VUV: 118.722
		WST: 2.8
		XAF: 606.502765
		XAG: 0.0295431
		XAU: 0.00036745
		XCD: 2.70255
		XDR: 0.749896
		XOF: 606.502765
		XPD: 0.00094831
		XPF: 110.335025
		XPT: 0.00099611
		YER: 250.375001
		ZAR: 17.616125
		ZMW: 26.669432
		ZWL: 322
	}
}